backend/flask/requirements.txt
# Install dependencies with pip install -r requirements.txt
Flask>=3.1.2
python-dotenv>=1.1.1
requests>=2.32.5
supabase>=2.22.3
